// projects/evo-lib/src/lib/evo/devices-screen/screen/screen.enum.ts
/**
 * Типы данных для 'Операционной системы'.
 */
export enum OperatingSystemEnum {
    Windows = 'Windows',
    MacOS = 'MacOS',
    Linux = 'Linux',
    Android = 'Android',
    iOS = 'iOS',
    UNIX = 'UNIX',
    Unknown = 'Unknown'
}

/**
 * Типы данных для 'Браузера'.
 */
export enum BrowserEnum {
    Edge = 'Microsoft Edge',
    Chrome = 'Google Chrome',
    Firefox = 'Mozilla Firefox',
    Safari = 'Safari',
    Opera = 'Opera',
    InternetExplorer = 'Internet Explorer',
    Unknown = 'Unknown'
}

/**
 * Типы данных для 'Экрана'.
 */
export enum ScreenEnum {
    Desktop = 'desktop',
    Tablet = 'tablet',
    Phone = 'phone',
}

/**
 * Типы данных для 'расположение экрана'.
 */
export enum OrientationScreenEnum {
    Horizontal = 'horizontal',
    Vertical = 'vertical',
}